Refactored file structure, added events controller
